clean up helpers; everything on Messer proto
