+++ event-handlers.js
@@ -1,5 +1,4 @@
 const log = require("./log")
-const helpers = require("./helpers")
 const fbAssets = require("./fb-assets")
 
 /**
